Moving cdn update to be before lerna bootstrap
