+++ lerna.json
@@ -3,5 +3,5 @@
   "packages": [
     "packages/*"
   ],
-  "version": "3.0.0-alpha.16"
+  "version": "3.0.0-alpha.17"
 }
