+++ packages/workbox-build/package.json
@@ -1,6 +1,6 @@
 {
   "name": "workbox-build",
-  "version": "3.0.0-alpha.16",
+  "version": "3.0.0-alpha.17",
   "description": "A module that integrates into your build process, helping you generate a manifest of local files that workbox-sw should precache.",
   "keywords": [
     "workbox",
