+++ packages/workbox-build/src/cdn-details.json
@@ -2,5 +2,5 @@
   "origin": "https://storage.googleapis.com",
   "bucketName": "workbox-cdn",
   "releasesDir": "releases",
-  "latestVersion": "3.0.0-alpha.16"
+  "latestVersion": "3.0.0-alpha.17"
 }
