+++ packages/workbox-core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "workbox-core",
-  "version": "3.0.0-alpha.16",
+  "version": "3.0.0-alpha.17",
   "license": "Apache-2.0",
   "author": "Google's Web DevRel Team",
   "description": "This module is used by a number of the other workboxjs.org modules to share common code.",
